StageとStage1のメンバ変数とデストラクタを整理
