+++ DX22Base/Scene.cpp
@@ -54,6 +54,9 @@ CScene::CScene()
 =========================================== */
 CScene::~CScene()
 {
+	SAFE_DELETE(m_pCamera);
+	SAFE_DELETE(m_pBGM);
+	SAFE_DELETE(m_pBGMSpeaker);
 }
 
 /* ========================================
@@ -158,19 +161,4 @@ void CScene::Draw2d(float posX, float posY, float h, float w, Texture* pTexture)
 CCamera* CScene::GetCamera()
 {
 	return m_pCamera;
-}
-
-/* ========================================
-   サウンドファイル読み込み関数
-   -------------------------------------
-   内容：サウンドファイルの読み込み
-   -------------------------------------
-   引数1：無し
-   -------------------------------------
-   戻値：無し
-=========================================== */
-void CScene::LoadSound()
-{
-	m_pBGM = CSound::LoadSound("Assets/Sound/BGM/BGM_maou.mp3", true);		//BGMの読み込み
-	m_pSEHitHammer = CSound::LoadSound("Assets/Sound/SE/Smash.mp3");		//SEの読み込み
 }
