+++ DX22Base/SceneGame.h
@@ -86,7 +86,7 @@ private:
 
 	XAUDIO2_BUFFER* m_pBGM;							//BGMの音声データ
 	XAUDIO2_BUFFER* m_pSEHitHammer;					//SEの音声データ
-	IXAudio2SourceVoice* m_pSpeaker;				//BGMを聞き取る側
+	IXAudio2SourceVoice* m_pBGMSpeaker;				//BGMを聞き取る側
 	IXAudio2SourceVoice* m_pSEHitHammerSpeaker;		//SEを聞き取る側
 
 };
