+++ DX22Base/Stage1.h
@@ -34,18 +34,9 @@ public:
 	E_TYPE GetType() const override;	//自身の種類ゲッタ
 	E_TYPE GetNext() const override;	//次のシーンゲッタ
 private:
-	int m_nNum;			// ゲームスタート表示カウント用
-	float m_fSize;		// ゲームスタート表示のサイズ
-	float m_fResize;	// ゲームスタート表示のサイズ変更用
-	bool m_bStart;		// ゲームを開始させるか判定
-	const int* m_pPlayerHp;		// プレイヤーHP取得用
-	const int* m_pTimeCnt;		// 制限時間取得用
-
-	TPos2d<float> m_fUVPos;	// UV座標保存用
-	int m_nCntSwitch;	// アニメーション切り替えカウント用
-	int m_nCntW;		// 横カウント用
-	int m_nCntH;		// 縦カウント用
-	bool m_bStartSign;		// スタート合図開始フラ
+	
+private:
+	void LoadBGM();
 };	//ステージ1
 
 #endif	//!__STAGE1_H__
